Fix extra category icons with similar file name prefix and different size
In case we have an extra category in the project structure,
like default possible category called Unsorted.
For example, we have folder with default Unsorted category and two files
in it:
  /src/svg/Unsorted/live_badge_14.svg
  /src/svg/Unsorted/live_badge_18.svg

generateIcons() function will generate single file for them in ts
folder:
  /ts/Unsorted/live_badge.ts

With single component called Icon18LiveBadge.
And in the ts/index.ts file we will have two exports, one
of them with error because there is no Icon14LiveBadge in the file:
  export { Icon14LiveBadge } from './Unsorted/live_badge'; <-- error !!!
  export { Icon18LiveBadge } from './Unsorted/live_badge';

To fix this we add `size` postfix to the ts file name.
So, generateIcons() will generate two files per components:
  /ts/Unsorted/live_badge_14.ts with Icon14LiveBadge
  /ts/Unsorted/live_badge_18.ts with Icon18LiveBadge

As a result of this we will have a lot of files with postfix `_${size}`
in the ts file name, even though they have a dirname with size already.
  /ts/20/badge_20.ts
  /dist/20/badge_20.js